pyomnirepo/storage/pyproject.toml
[project]
name = "storage"
version = "0.1.0"
description = ""
authors = [{ name = "olatejulian", email = "julian.lf.olate@gmail.com" }]
dependencies = [
    "fastapi>=0.78.0",
    "SQLAlchemy>=1.4.36",
    "python-interface>=1.6.1",
    "python-dotenv>=0.20.0",
]
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
